feature: transaction fix: fix issue nominal uang

+++ db/migrations/20250608095900_contract_products.sql
@@ -6,10 +6,10 @@ CREATE TABLE IF NOT EXISTS contract_products (
     kode_barang CHAR(26) NOT NULL, -- disamakan dengan products.kode
     jumlah INTEGER NOT NULL DEFAULT 0,
     satuan CHAR(50) NOT NULL,
-    harga_satuan INTEGER NOT NULL DEFAULT 0,
+    harga_satuan BIGINT NOT NULL DEFAULT 0,
     kode_mata_uang CHAR(50) NOT NULL,
-    nilai_barang_fog INTEGER NOT NULL DEFAULT 0,
-    nilai_barang_rp INTEGER NOT NULL DEFAULT 0,
+    nilai_barang_fog BIGINT NOT NULL DEFAULT 0,
+    nilai_barang_rp BIGINT NOT NULL DEFAULT 0,
     created_at TIMESTAMP WITH TIME ZONE DEFAULT CURRENT_TIMESTAMP NOT NULL,
     updated_at TIMESTAMP WITH TIME ZONE DEFAULT CURRENT_TIMESTAMP NOT NULL,
     deleted_at TIMESTAMP WITH TIME ZONE,
